Add menu scene with New Game option
